Merge puppeteer sourcemaps internally
There was a very strange thing happening where the coverage map merge
function in istanbul would sometimes fail if the same file was passed
in from different coverage reports.

There is still something wonky going on with the branches and functions
reporting from istanbul cause it seems to always return 100% when using
the puppeteer coverage. Probably this is an issue with v8-to-istanbul
or converting to v8.