Use default puppeteer timeout
Hardcoding this timeout can cause issues when your code builds become
large or your computer is under load. That is because this timeout is
not testing the actual load times of your pages, but rather the time it
takes for the test bundles to get built. Since it runs rollup on the
fly at the time of the request, it could take a few seconds just to get
the initial page to load. The issue is worse if you dynamically inject
other scripts into your tests at runtime.